Move championship seeds into seeds and fix bug
* Move champiohship seeds into seeds file since it is basic site details
* Fix bug where fantasy team names in seed file exceeded 16 chars
* Closes #139
