+++ priv/repo/csv_seed_data/fantasy_teams.csv
@@ -7,7 +7,7 @@ Wild Cats,6,0,0,2
 Black Mambas,1,0,0,1
 Bengal Tigers,2,0,0,1
 Huskies,12,0,0,1
-Rosie's Riveters,3,0,0,1
+Riveters,3,0,0,1
 Junior Mints,10,0,0,2
 She-Unit,5,0,0,1
 Black Widows,9,0,0,1
@@ -16,14 +16,14 @@ Bare Exposure,9,0,0,2
 Pink Cougars,7,0,0,1
 End Zone Divas,11,0,0,1
 Girl Power,11,0,0,2
-Victorious Secret,4,0,0,1
+Victorious S,4,0,0,1
 Air Farce,1,0,0,2
-Team Jagermeister,13,0,0,2
+Jagermeister,13,0,0,2
 Hoosier Daddys,8,0,0,2
 Ocho Stinko,5,0,0,2
 No Punt Intended,8,0,0,1
 Loose Ends,2,0,0,2
-Little Jerry Seinfelds,4,0,0,2
+Jerry Seinfelds,4,0,0,2
 Deep Threat,3,0,0,2
-Back that Pass Up,12,0,0,2
+Pass Up,12,0,0,2
 Fumble This,14,0,0,2
